+++ docs/source/transforms/convert_posargs_to_args.py
@@ -1,0 +1,6 @@
+def name(p1, p2, /, p_or_kw, *, kw): pass
+def name(p1, p2=None, /, p_or_kw=None, *, kw): pass
+def name(p1, p2=None, /, *, kw): pass
+def name(p1, p2=None, /): pass
+def name(p1, p2, /, p_or_kw): pass
+def name(p1, p2, /): pass