+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.2.2] - 2019-11-16
+## [2.3.0] - Unreleased
+
+### Added
+- Optional source transform:
+    - convert positional arguments to normal arguments, enabled by default
 
 ### Fixed
 - Unnecessary spaces after ',' in tuple values
+- Removing annotations for positional-only arguments (Thanks [luk3yx](https://github.com/luk3yx)!)
 
 ## [2.2.1] - 2019-11-03
 
@@ -76,6 +81,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - pyminify command
 
 
+[2.3.0]: https://github.com/dflook/python-minifier/compare/2.2.1...2.3.0
 [2.2.1]: https://github.com/dflook/python-minifier/compare/2.2.0...2.2.1
 [2.2.0]: https://github.com/dflook/python-minifier/compare/2.1.2...2.2.0
 [2.1.2]: https://github.com/dflook/python-minifier/compare/2.1.1...2.1.2